Add files via upload
Z-score function created to eliminate the age-related immune changes between age-groups. The last function in the script is used in the cross-validation step where the standard deviation and mean are calculated from the healthy controls of the training set and applied to the patients in the test set by the z-score.